requirements.txt
gspread==3.6.0
google-auth==1.16.0
